Improve content driven mode, article css done
