+++ assets/style/main.css
@@ -74,3 +74,17 @@ html, body, #__nuxt, #__layout {
     @apply tracking-wider;
   }
 }
+
+.writing {
+  @apply prose antialiased text-gray-700/90 dark:text-gray-400/80 leading-relaxed;
+  @apply prose-h1:text-2xl prose-h1:mb-0 sm:prose-h1:text-3xl prose-h1:text-zinc-900 dark:prose-h1:text-zinc-100 prose-h1:italic prose-h1:font-[600];
+  @apply prose-h2:text-xl sm:prose-h2:text-2xl prose-h2:text-zinc-900 dark:prose-h2:text-zinc-100;
+  @apply prose-h3:text-lg sm:prose-h3:text-xl prose-h3:text-zinc-900 dark:prose-h3:text-zinc-100;
+  @apply prose-h4:text-base sm:prose-h4:text-lg prose-h4:text-zinc-900 dark:prose-h4:text-zinc-100;
+  @apply prose-h5:text-base prose-h5:text-zinc-900 dark:prose-h5:text-zinc-100;
+  @apply prose-a:text-blue-500 dark:prose-a:text-blue-500 prose-a:underline;
+  @apply prose-pre:text-sm sm:prose-pre:text-base prose-pre:text-zinc-900 dark:prose-pre:text-zinc-100;
+  @apply prose-code:text-zinc-900 dark:prose-code:text-zinc-100;
+  @apply prose-strong:text-gray-900 dark:prose-strong:text-gray-100 prose-strong:font-[500];
+  @apply prose-p:m-0;
+}
