+++ content/en/2.works.md
@@ -1,3 +1,7 @@
+---
+title: Projects
+---
+
 ::works
 #title
 Projects
