+++ content/en/3.writing.md
@@ -1,3 +1,7 @@
+---
+title: Writing
+---
+
 ::writing
 #title
 Writing
