+++ content/en/5.contact.md
@@ -1,7 +1,11 @@
+---
+title: Contact
+---
+
 ::contact
 #title
-Contactez moi
+Contact Me
 
 #subtitle
-Discutons de votre idée, projet ou de tout autre chose !
+Let's talk about your ideas, projects or anything else
 ::
