+++ content/fr/2.works.md
@@ -1,3 +1,7 @@
+---
+title: Projets
+---
+
 ::works
 #title
 Projets
