+++ content/fr/3.writing.md
@@ -1,3 +1,7 @@
+---
+title: Articles
+---
+
 ::writing
 #title
 Articles
