+++ content/fr/4.about.md
@@ -1,3 +1,7 @@
+---
+title: A propos
+---
+
 ::about
 #title
 A propos
