+++ content/fr/5.contact.md
@@ -1,7 +1,11 @@
+---
+title: Contact
+---
+
 ::contact
 #title
-Contact me
+Contactez moi
 
 #subtitle
-Let's talk about your ideas, projects or anything else
+Discutons de votre idée, projet ou de tout autre chose !
 ::
