--- locales/en/about.json
@@ -1,5 +1,0 @@
-{
-  "title": "About",
-  "description": "Some information about me",
-  "experience": "Experience"
-}
