+++ locales/en/writing.json
@@ -10,8 +10,6 @@
   "copy_link" : "Copy link",
   "search_article" : "Search article",
   "blog" : "Blog",
-  "quotes" : "Quotes",
-  "empty_quotes" : "No quotes yet, check back soon!",
   "unknown_author" : "Unknown author",
   "show_search" : "Search",
   "hide_search" : "Hide search"
