--- locales/fr/about.json
@@ -1,5 +1,0 @@
-{
-  "title": "A Propos",
-  "description": "Quelques informations sur moi",
-  "experience": "Expérience"
-}
