+++ locales/fr/contact.json
@@ -7,7 +7,6 @@
   "subject": "Sujet de votre message",
   "message": "Message",
   "submit": "Envoyer",
-  "average_response_time" : "Je réponds en moyenne entre 2 et 5 heures",
   "success": "Votre message a été envoyé avec succès",
   "error": "Oups, une erreur s'est produite lors de l'envoi de votre message :(",
   "budget": "Budget",
