+++ locales/fr/writing.json
@@ -10,8 +10,6 @@
   "copy_link" : "Copier le lien",
   "search_article" : "Rechercher un article",
   "blog" : "Blog",
-  "quotes" : "Citations",
-  "empty_quotes" : "Aucune citation pour le moment, mais ça ne saurait tarder !",
   "unknown_author" : "Auteur inconnu",
   "show_search" : "Recherche",
   "hide_search" : "Masquer la recherche"
